+++ packages/patternfly-4/react-docs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.14.29](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.14.28...@patternfly/react-docs@4.14.29) (2019-10-18)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
 ## [4.14.28](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.14.27...@patternfly/react-docs@4.14.28) (2019-10-17)
 
 **Note:** Version bump only for package @patternfly/react-docs