+++ packages/patternfly-4/react-integration/demo-app-ts/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.13](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@3.7.12...demo-app-ts@3.7.13) (2019-10-18)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [3.7.12](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@3.7.11...demo-app-ts@3.7.12) (2019-10-17)
+
+
+### Bug Fixes
+
+* **table:** set height auto modifier on table row ([#3133](https://github.com/patternfly/patternfly-react/issues/3133)) ([6e89ebe](https://github.com/patternfly/patternfly-react/commit/6e89ebe))
+
+
+
+
+
 ## [3.7.11](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@3.7.10...demo-app-ts@3.7.11) (2019-10-17)
 
 **Note:** Version bump only for package demo-app-ts