+++ packages/patternfly-4/react-topology/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.9.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.9.6...@patternfly/react-topology@2.9.7) (2019-10-18)
+
+**Note:** Version bump only for package @patternfly/react-topology
+
+
+
+
+
 ## [2.9.6](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-topology@2.9.5...@patternfly/react-topology@2.9.6) (2019-10-17)
 
 **Note:** Version bump only for package @patternfly/react-topology